Cleaned up includes properly and added concept of scene to abstract away the 3D level and make it easier to create different ones. Now to 2D I think!

+++ RPG_Project/include/Camera.h
@@ -2,15 +2,15 @@
 
 #include <glm/glm.hpp>
 
-const float CAM_DEFAULT_YAW = -90.0f;
-const float CAM_DEFAULT_PITCH = 0.0f;
-const float CAM_DEFAULT_SPEED = 2.5f;
-const float CAM_DEFAULT_SENSITIVITY = 0.1f;
-const float CAM_DEFAULT_FOV = 45.0f;
+const double CAM_DEFAULT_YAW = -90.0f;
+const double CAM_DEFAULT_PITCH = 0.0f;
+const double CAM_DEFAULT_SPEED = 2.5f;
+const double CAM_DEFAULT_SENSITIVITY = 0.1f;
+const double CAM_DEFAULT_FOV = 45.0f;
 
-const float CAM_MAX_PITCH = 89.999f;
-const float CAM_MIN_FOV = 1.0f;
-const float CAM_MAX_FOV = 90.0f;
+const double CAM_MAX_PITCH = 89.999f;
+const double CAM_MIN_FOV = 1.0f;
+const double CAM_MAX_FOV = 90.0f;
 
 enum class CameraDirection {
     FORWARD,
@@ -22,11 +22,12 @@ enum class CameraDirection {
 };
 
 class Camera {
+
 public:
     glm::vec3 position;
-    float movementSpeed;
-    float mouseSensitivity;
-    float fov;
+    double movementSpeed;
+    double mouseSensitivity;
+    double fov;
 
 private:
     glm::vec3 front;
@@ -34,18 +35,18 @@ private:
     glm::vec3 right;
     glm::vec3 worldUp;
 
-    float yaw;
-    float pitch;
+    double yaw;
+    double pitch;
 
 public:
-    Camera(glm::vec3 position = glm::vec3(0, 0, 0), glm::vec3 up = glm::vec3(0, 1, 0), float yaw = CAM_DEFAULT_YAW, float pitch = CAM_DEFAULT_PITCH);
+    Camera(glm::vec3 position = glm::vec3(0, 0, 0), glm::vec3 up = glm::vec3(0, 1, 0), double yaw = CAM_DEFAULT_YAW, double pitch = CAM_DEFAULT_PITCH);
 
     glm::mat4 getViewMatrix();
-    glm::mat4 getPerspectiveProjectionMatrix(float aspectRatio, float zNear, float zFar);
+    glm::mat4 getPerspectiveProjectionMatrix(double aspectRatio, double zNear, double zFar);
 
-    void processKeyboardInput(CameraDirection direction, float deltaTime);
-    void processMouseMovement(float xOffset, float yOffset);
-    void processMouseScroll(float yOffset);
+    void processKeyboardInput(CameraDirection direction, double deltaTime);
+    void processMouseMovement(double xOffset, double yOffset);
+    void processMouseScroll(double yOffset);
 
 private:
     void updateCameraVectors();
